fix(regressor): keep predictions exact for targets far from zero
A target whose mean is large relative to its std collapsed the raw-space
bar distribution's float32 borders, so fit raised an AssertionError or,
when the outer buckets survived, predicted from mostly zero-width buckets.
Build the raw-space criterion in float64 and decode around a float64
offset so the float32 decode only resolves the target's spread.

+++ src/tabpfn/finetuning/data_util.py
@@ -455,9 +455,9 @@ class DatasetCollectionWithPreprocessing(torch.utils.data.Dataset):
             y_test_standardized = (y_test_raw - train_mean) / train_std
             y_train_standardized = (y_train_raw - train_mean) / train_std
             raw_space_bardist_ = FullSupportBarDistribution(
-                znorm_space_bardist_.borders * train_std
+                znorm_space_bardist_.borders.double() * train_std
                 + train_mean  # Inverse normalization back to raw space
-            ).float()
+            )
             y_train = y_train_standardized
         else:
             y_train = y_train_raw
